backend/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
sqlalchemy==2.0.38
pydantic==2.10.6
python-jose==3.4.0
passlib==1.7.4
python-multipart==0.0.20
bcrypt==4.3.0
pytest==8.3.5
email-validator==2.2.0
httpx==0.27.0
pillow==10.3.0